refactor(controller): finish push

+++ CrossCutting/DataTransferObjects/WorkingTimeDto.cs
@@ -4,7 +4,7 @@ public class WorkingTimeDto
 {
     public int Id { get; set; }
     public DateTime? Date { get; set; }
-    public decimal TimeEntry { get; set; }
+    public double TimeEntry { get; set; }
     public int ProjectId { get; set; }
     public int CategoryId { get; set; }
     public int EmployeeId { get; set; }
